REFERENCE: QGA-61

+++ controllers/userController.js
@@ -1,7 +1,7 @@
 import { getUsers, updateUser, deleteUser } from '../models/UserModel';
 import AppError from '../errors/AppError';
 import { accessLevelCheck } from '../helpers/accessLevelCheck';
-const bcrypt = require('bcrypt');
+import bcrypt from 'bcrypt';
 
 const logger = require('../utils/logger')('logController');
 
@@ -33,33 +33,16 @@ const getAllUsers = async (req, res, next) => {
 const updateUserById = async (req, res, next) => {
   try {
     const id = req.params.userId;
-    if (accessLevelCheck(req.user.level, 'admin')) {
+    if (req.user.id === id || accessLevelCheck(req.user.level, 'admin')) {
       const { level, hashedPassword, name, surname, email, dateOfBirth } = req.body;
       let userUpdate = {};
+
       if (hashedPassword) {
         const reshedPassword = await bcrypt.hash(hashedPassword, parseInt(process.env.PASSWORD_HASHING_ROUNDS, 10));
         userUpdate.hashedPassword = reshedPassword;
       }
-      level ? (userUpdate.level = level) : null;
-      name ? (userUpdate.name = name) : null;
-      surname ? (userUpdate.surname = surname) : null;
-      email ? (userUpdate.email = email) : null;
-      dateOfBirth ? (userUpdate.dateOfBirth = dateOfBirth) : null;
-
-      const updatedUser = await updateUser(id, userUpdate);
-      res.status(200).send({
-        payload: { user: updatedUser },
-      });
-      return;
-    }
-
-    if (req.user.id === id) {
-      const { hashedPassword, name, surname, email, dateOfBirth } = req.body;
-      let userUpdate = {};
-
-      if (hashedPassword) {
-        const reshedPassword = await bcrypt.hash(hashedPassword, parseInt(process.env.PASSWORD_HASHING_ROUNDS, 10));
-        userUpdate.hashedPassword = reshedPassword;
+      if (req.user.level === 'admin') {
+        level ? (userUpdate.level = level) : null;
       }
       name ? (userUpdate.name = name) : null;
       surname ? (userUpdate.surname = surname) : null;
@@ -74,7 +57,6 @@ const updateUserById = async (req, res, next) => {
     } else {
       throw new AppError('User can only change onw data');
     }
-    next();
   } catch (error) {
     next(new AppError(error.message));
   }
@@ -83,7 +65,7 @@ const updateUserById = async (req, res, next) => {
 const deleteUserById = async (req, res, next) => {
   try {
     const id = req.params.userId;
-    if (accessLevelCheck(req.user.level, 'admin')) {
+    if (accessLevelCheck(req.user.level, 'admin') && req.user.id !== id) {
       const deletedUser = await deleteUser(id);
       if (deletedUser) {
         res.status(200).send({
@@ -92,13 +74,13 @@ const deleteUserById = async (req, res, next) => {
           },
         });
       } else {
-        throw new AppError('user not found');
+        throw new AppError('User not found', 400);
       }
     } else {
-      throw new AppError('Only admin can delete the user');
+      throw new AppError("Only admin can delete the user and admin can't delete himself", 400);
     }
   } catch (error) {
-    next(error instanceof AppError ? error : new AppError(error.message));
+    next(error instanceof AppError ? error : new AppError(error.message, 400));
   }
 };
